+++ client_code/utils/_timed.py
@@ -6,36 +6,45 @@
 # This software is published at https://github.com/anvilistas/anvil-extras
 
 
+import sys
 from functools import wraps
-from time import gmtime, strftime, time
+from time import time
+
+from .. import logging as logging
 
 __version__ = "1.9.0"
+
+LOGGER = logging.Logger(name="timing", format="{datetime:%Y-%m-%d %H:%M:%S}: {msg}")
 
 
 def _signature(func, args, kwargs):
     """Text representation of a function's signature"""
-    arguments = [str(a) for a in args]
-    arguments.extend([f"{key}={value}" for key, value in kwargs.items()])
-    return f"{func.__name__}({','.join(arguments)})"
+    sig = [repr(a) for a in args]
+    sig.extend(f"{key}={value!r}" for key, value in kwargs.items())
+    sig = ",".join(sig)
+    if len(sig) > 40:
+        sig = sig[:40] + "..."
+
+    return f"{func.__name__}({sig})"
 
 
-def _timestamp(seconds):
-    """Text representation of a unix timestamp"""
-    return strftime("%Y-%m-%d %H:%M:%S", gmtime(seconds))
-
-
-def timed(func):
+def timed(_func=None, *, logger=LOGGER, level=logging.INFO):
     """A decorator to time the execution of a function"""
+    if _func is None:
+        return lambda _func: timed(_func, logger=logger, level=level)
 
-    @wraps(func)
+    @wraps(_func)
     def wrapper(*args, **kwargs):
-        signature = _signature(func, args, kwargs)
+        signature = _signature(_func, args, kwargs)
+        logger.log(msg=f"{signature} called", level=level)
         started_at = time()
-        print(f"{_timestamp(started_at)} {signature} called")
-        result = func(*args, **kwargs)
+        result = _func(*args, **kwargs)
         finished_at = time()
         duration = f"{(finished_at - started_at):.2f}s"
-        print(f"{_timestamp(finished_at)} {signature} completed({duration})")
+        logger.log(
+            msg=f"{signature} completed({duration})",
+            level=level,
+        )
         return result
 
     return wrapper